requirements.txt
RPi.GPIO (for raspberry pi)
numpy
tensorflow
opencv-python
microsoft tools needed (windows machines)